Create separate btc and stx rbf components
