+++ src/app/screens/speedUpTransaction/index.styled.ts
@@ -1,13 +1,4 @@
-import ActionButton from '@components/button';
-import { Faders } from '@phosphor-icons/react';
 import styled from 'styled-components';
-
-export const Title = styled.h1((props) => ({
-  ...props.theme.typography.headline_s,
-  color: props.theme.colors.white_0,
-  marginTop: props.theme.spacing(8),
-  marginBottom: props.theme.spacing(8),
-}));
 
 export const LoaderContainer = styled.div({
   display: 'flex',
@@ -16,116 +7,13 @@ export const LoaderContainer = styled.div({
   height: 'inherit',
 });
 
-export const Container = styled.div((props) => ({
-  display: 'flex',
-  flexDirection: 'column',
-  paddingLeft: props.theme.spacing(8),
-  paddingRight: props.theme.spacing(8),
-  ...props.theme.scrollbar,
-}));
-
-export const DetailText = styled.span((props) => ({
-  ...props.theme.typography.body_m,
-  color: props.theme.colors.white_200,
-  marginBottom: props.theme.spacing(4),
-}));
-
-export const HighlightedText = styled.span((props) => ({
-  ...props.theme.typography.body_medium_m,
-  color: props.theme.colors.white_0,
-}));
-
-export const ButtonContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  margin-top: ${(props) => props.theme.spacing(6)}px;
-  gap: ${(props) => props.theme.spacing(4)}px;
-`;
-
-export const FeeButton = styled.button<{
-  isSelected: boolean;
-  centered?: boolean;
-}>((props) => ({
-  ...props.theme.body_medium_m,
-  textAlign: 'left',
-  color: props.theme.colors.white_0,
-  backgroundColor: `${props.isSelected ? props.theme.colors.elevation6_600 : 'transparent'}`,
-  border: `1px solid ${
-    props.isSelected ? props.theme.colors.white_800 : props.theme.colors.white_850
-  }`,
-  borderRadius: props.theme.radius(2),
-  height: 'auto',
-  display: 'flex',
-  justifyContent: 'space-between',
-  alignItems: props.centered ? 'center' : 'flex-start',
-  transition: 'background-color 0.1s ease-in-out, border 0.1s ease-in-out',
-  padding: props.theme.spacing(8),
-  paddingTop: props.theme.spacing(6),
-  paddingBottom: props.theme.spacing(6),
-  ':not(:disabled):hover': {
-    borderColor: props.theme.colors.white_800,
-  },
-  ':disabled': {
-    cursor: 'not-allowed',
-    color: props.theme.colors.white_400,
-    div: {
-      color: 'inherit',
-    },
-    svg: {
-      fill: props.theme.colors.white_600,
-    },
-  },
-}));
-
-export const ControlsContainer = styled.div`
-  display: flex;
-  column-gap: 12px;
-  margin: 38px 0px 40px;
-`;
-
-export const CustomFeeIcon = styled(Faders)({
-  transform: 'rotate(90deg)',
-});
-
-export const FeeButtonLeft = styled.div((props) => ({
-  display: 'flex',
-  alignItems: 'center',
-  gap: props.theme.spacing(6),
-}));
-
-export const FeeButtonRight = styled.div({
-  textAlign: 'right',
-});
-
-export const SecondaryText = styled.div<{
-  alignRight?: boolean;
-}>((props) => ({
-  ...props.theme.typography.body_medium_s,
-  color: props.theme.colors.white_200,
-  marginTop: props.theme.spacing(2),
-  textAlign: props.alignRight ? 'right' : 'left',
-}));
-
-export const StyledActionButton = styled(ActionButton)((props) => ({
-  'div, h1': {
-    ...props.theme.typography.body_medium_m,
-  },
-}));
-
-export const WarningText = styled.span((props) => ({
-  ...props.theme.typography.body_medium_s,
-  display: 'block',
-  color: props.theme.colors.danger_light,
-  marginTop: props.theme.spacing(2),
-}));
-
 export const SuccessActionsContainer = styled.div((props) => ({
   width: '100%',
   display: 'flex',
   flexDirection: 'column',
-  gap: props.theme.spacing(6),
-  paddingLeft: props.theme.spacing(8),
-  paddingRight: props.theme.spacing(8),
-  marginBottom: props.theme.spacing(20),
-  marginTop: props.theme.spacing(20),
+  gap: props.theme.space.s,
+  paddingLeft: props.theme.space.m,
+  paddingRight: props.theme.space.m,
+  marginBottom: props.theme.space.xxl,
+  marginTop: props.theme.space.xxl,
 }));
